Move UserReviews component to ReviewsPage folder. Remove unused scss file
